+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Call an other (external) URL/endpoint from a hook
 - Native Sentry message support
 - Support for (Inoreader filter) RSS news item notification
+- Support for Freshping monitoring updates
 
 ### Changed
 - Use ruff to lint and fix codebase